src/main/java/com/coding/challenge/transactions/service/AccountService.java
package com.coding.challenge.transactions.service;

import com.coding.challenge.transactions.entity.Account;
import com.coding.challenge.transactions.repository.AccountRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class AccountService {

    @Autowired
    private AccountRepository accountRepository;

    public List<Account> getCustomerAccounts(String customerId) {
        return accountRepository.findByCustomerId(customerId);
    }

}
